refactor: rename shared/wake_lifecycle to shared/pending_scan_lifecycle
Mechanical rename to align module name with the new cron-based
mechanism. File content preserved verbatim (count_active_tasks,
has_same_teammate_continuation, _lifecycle_relevant remain
mechanism-agnostic lifecycle helpers).

Renamed:
- pact-plugin/hooks/shared/wake_lifecycle.py -> shared/pending_scan_lifecycle.py
- pact-plugin/tests/test_inbox_wake_lifecycle_helper.py ->
  pact-plugin/tests/test_pending_scan_lifecycle_helper.py

Import updates across all 5 consumer sites (atomic rename per
PR #723 cycle-1 multi-file revert discipline):
- pact-plugin/hooks/wake_lifecycle_emitter.py (line 134)
- pact-plugin/hooks/session_init.py (line 84)
- pact-plugin/tests/test_pending_scan_lifecycle_helper.py (line 28)
- pact-plugin/tests/test_intentional_wait.py (lines 685, 1213, 1255)
- pact-plugin/tests/test_has_same_teammate_continuation.py (line 24)

The `wake_lifecycle_emitter.py` and `tests/fixtures/wake_lifecycle/`
paths are intentionally NOT renamed in this commit (emitter retains
its name; fixtures directory is test-engineer-owned scope).

Refs #726
